create generic file for resolver methods that are shared between darwin and non-darwin unix platforms
